refactor: 불필요 모듈(bp-dto, bp-core-web) 제거 및 통합
