+++ bp-app-api/build.gradle
@@ -5,17 +5,15 @@ plugins {
 }
 
 dependencies {
-//    implementation project(':bp-domain-rdb')
     implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    implementation project(':bp-domain-rdb')
     implementation project(':bp-kafka-event-publisher')
-    implementation project(':bp-dto')
     implementation project(':bp-utils')
-    implementation project(':bp-core-web')
 
     implementation 'org.apache.commons:commons-lang3:3.14.0'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    testImplementation (testFixtures(project(':bp-dto')))
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.junit.jupiter:junit-jupiter-api'
