+++ bp-app-api/src/main/java/com/beautify_project/bp_app_api/dto/shop/ImageFiles.java
@@ -1,4 +1,4 @@
-package com.beautify_project.bp_dto.shop;
+package com.beautify_project.bp_app_api.dto.shop;
 
 import java.util.ArrayList;
 import java.util.List;
