+++ bp-app-api/src/main/java/com/beautify_project/bp_app_api/enumeration/ShopSearchType.java
@@ -1,6 +1,6 @@
-package com.beautify_project.bp_dto.shop.enumeration;
+package com.beautify_project.bp_app_api.enumeration;
 
-import com.beautify_project.bp_dto.exception.EnumMismatchException;
+import com.beautify_project.bp_app_api.exception.EnumMismatchException;
 import org.apache.commons.lang3.StringUtils;
 
 public enum ShopSearchType {
