+++ bp-app-api/src/main/java/com/beautify_project/bp_app_api/service/ShopService.java
@@ -1,9 +1,9 @@
 package com.beautify_project.bp_app_api.service;
 
-import com.beautify_project.bp_dto.common.response.ResponseMessage;
-import com.beautify_project.bp_dto.shop.ImageFiles;
-import com.beautify_project.bp_dto.shop.ShopFindListRequestParameters;
-import com.beautify_project.bp_dto.shop.ShopRegistrationRequest;
+import com.beautify_project.bp_app_api.dto.common.ResponseMessage;
+import com.beautify_project.bp_app_api.dto.shop.ImageFiles;
+import com.beautify_project.bp_app_api.dto.shop.ShopFindListRequestParameters;
+import com.beautify_project.bp_app_api.dto.shop.ShopRegistrationRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
