--- bp-dto/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-    id 'java-test-fixtures'
-    id 'org.springframework.boot' version '3.3.5'
-    id 'io.spring.dependency-management' version '1.1.6'
-    id 'io.freefair.lombok' version '8.11'
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.apache.commons:commons-lang3:3.14.0'
-
-    testFixturesImplementation 'org.springframework.boot:spring-boot-starter-web'
-    testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testFixturesImplementation 'org.apache.commons:commons-lang3:3.14.0'
-}
-
-bootJar {
-    enabled = false
-}
