--- bp-dto/src/main/java/com/beautify_project/bp_dto/shop/ShopFindListRequestParameters.java
@@ -1,9 +1,0 @@
-package com.beautify_project.bp_dto.shop;
-
-import com.beautify_project.bp_dto.shop.enumeration.OrderType;
-import com.beautify_project.bp_dto.shop.enumeration.ShopSearchType;
-
-public record ShopFindListRequestParameters(ShopSearchType searchType, Integer page, Integer count,
-                                            OrderType orderType) {
-
-}
