+++ bp-kafka-event-consumer/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+spring.application.name=bp-kafka-event-consumer
