+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'beautify-project'
-include ':bp-app-api', ':bp-app-batch', ':bp-core-web', ':bp-dto', ':bp-dto',
-        ':bp-kafka-event-publisher', ':bp-utils', ':bp-domain-rdb'
+include ':bp-app-api', ':bp-app-batch', ':bp-domain-rdb', ':bp-kafka-event-publisher',
+        ':bp-kafka-event-consumer', ':bp-utils'
